// src/types/profile.ts
export interface UserProfile {
    user_id: string;
    username?: string;
    full_name?: string;
    bio?: string;
    profile_picture_url?: string;
    hero_banner_url?: string;
    youtube_url?: string;
    instagram_url?: string;
    measurement_system?: string;
    privacy_setting?: string;
    created_at?: string;
    updated_at?: string;
} 